magnete reworked Magnete sind noch buggy und fangen an an den Wänden hoch zu laufen & realistische Bedingungne für start geben kein ansatz weise gutes ergebnis

+++ src/com/company/Line.java
@@ -1,47 +1,47 @@
 package com.company;
 public class Line {
 
-    int x0;     //Startpunkt der Linie
-    int x1;
+    double x0;     //Startpunkt der Linie
+    double x1;
 
-    int y0;     //Endpunkt der Linie
-    int y1;
+    double y0;     //Endpunkt der Linie
+    double y1;
 
 
-    public Line(int x0, int y0, int x1, int y1){
+    public Line(double x0, double y0, double x1, double y1){
         this.x0 = x0;
         this.x1 = x1;
         this.y0 = y0;
         this.y1 = y1;
     }
 
-    public int getX0(){ return x0; }
+    public double getX0(){ return x0; }
 
-    public int getX1(){ return x1; }
+    public double getX1(){ return x1; }
 
-    public int getY0(){
+    public double getY0(){
         return y0;
     }
 
-    public int getY1(){
+    public double getY1(){
         return y1;
     }
 
 
 
-    public void setX0(int x0) { this.x0 = x0; }
+    public void setX0(double x0) { this.x0 = x0; }
 
-    public void setX1(int x1) { this.x1 = x1; }
+    public void setX1(double x1) { this.x1 = x1; }
 
-    public void setY0(int y0) { this.y0 = y0; }
+    public void setY0(double y0) { this.y0 = y0; }
 
-    public void setY1(int y1) { this.y1 = y1; }
+    public void setY1(double y1) { this.y1 = y1; }
 
-    public void setP1(int x0, int y0){
+    public void setP1(double x0, double y0){
         this.x0 = x0 ;
         this.y0 = y0;
     }
-    public void setP2(int x1, int y1){
+    public void setP2(double x1, double y1){
         this.x1 = x1 ;
         this.y1 = y1;
     }
